feat: implement admin-only access control for account management routes and add reset password functionality

+++ Coolking_Server/src/controllers/account.controller.js
@@ -1,9 +1,16 @@
 const accountRepo = require('../repositories/account.repo');
 const tokenRepo = require('../repositories/token.repo');
+const jwtUtils = require('../utils/jwt.utils');
 
 // GET /accounts
 exports.getAllAccounts = async (req, res) => {
 	try {
+		const authHeader = req.headers['authorization'];
+		const token = authHeader && authHeader.split(' ')[1];
+		const decoded = jwtUtils.verifyAccessToken(token);
+		if (!decoded || decoded.role !== 'ADMIN') {
+			return res.status(403).json({ message: 'Forbidden' });
+		}
 		const accounts = await accountRepo.getAllAccounts_paging(req.query.page || 1, req.query.pagesize || 10);
 		res.json(accounts);
 	} catch (err) {
@@ -25,6 +32,12 @@ exports.getAccountById = async (req, res) => {
 // POST /accounts
 exports.createAccount = async (req, res) => {
 	try {
+		const authHeader = req.headers['authorization'];
+		const token = authHeader && authHeader.split(' ')[1];
+		const decoded = jwtUtils.verifyAccessToken(token);
+		if (!decoded || decoded.role !== 'ADMIN') {
+			return res.status(403).json({ message: 'Forbidden' });
+		}
 		if (!req.body.password ||req.body.password.trim() === '') {
 			req.body.password = '12345678';
 		}
@@ -36,8 +49,14 @@ exports.createAccount = async (req, res) => {
 };
 
 // PUT /accounts/:id
-exports.updateAccount = async (req, res) => {
+exports.updateAccount4Admin = async (req, res) => {
 	try {
+		const authHeader = req.headers['authorization'];
+		const token = authHeader && authHeader.split(' ')[1];
+		const decoded = jwtUtils.verifyAccessToken(token);
+		if (!decoded || decoded.role !== 'ADMIN') {
+			return res.status(403).json({ message: 'Forbidden' });
+		}
 		const updatedAccount = await accountRepo.updateAccount(req.params.id, req.body);
 		if (!updatedAccount) return res.status(404).json({ message: 'Account not found' });
 		res.status(200).json({ message: 'Account updated successfully' });
@@ -49,6 +68,12 @@ exports.updateAccount = async (req, res) => {
 // DELETE /accounts/:id
 exports.deleteAccount = async (req, res) => {
 	try {
+		const authHeader = req.headers['authorization'];
+		const token = authHeader && authHeader.split(' ')[1];
+		const decoded = jwtUtils.verifyAccessToken(token);
+		if (!decoded || decoded.role !== 'ADMIN') {
+			return res.status(403).json({ message: 'Forbidden' });
+		}
 		const deletedAccount = await accountRepo.deleteAccount(req.params.id);
 		if (!deletedAccount) return res.status(404).json({ message: 'Account not found' });
 		// Xóa token liên quan (nếu có)
@@ -59,11 +84,30 @@ exports.deleteAccount = async (req, res) => {
 	}
 };
 
+// POST /accounts/reset-password/:id
+exports.resetPassword4Admin = async (req, res) => {
+	try {
+		const authHeader = req.headers['authorization'];
+		const token = authHeader && authHeader.split(' ')[1];
+		const decoded = jwtUtils.verifyAccessToken(token);
+		if (!decoded || decoded.role !== 'ADMIN') {
+			return res.status(403).json({ message: 'Forbidden' });
+		}
+		const result = await accountRepo.resetPassword4Admin(req.params.id);
+		res.status(200).json({ message: 'Reset password successfully', newPassword: result });
+	} catch (err) {
+		res.status(500).json({ message: err.message });
+	}
+};
+
 // POST /accounts/change-password
 exports.changePassword = async (req, res) => {
 	try {
-		const { user_id, oldPassword, newPassword } = req.body;
-		await accountRepo.changePassword(user_id, oldPassword, newPassword);
+		const authHeader = req.headers['authorization'];
+		const token = authHeader && authHeader.split(' ')[1];
+		const decoded = jwtUtils.verifyAccessToken(token);
+		const {oldPassword, newPassword } = req.body;
+		await accountRepo.changePassword(decoded.user_id, oldPassword, newPassword);
 		res.status(200).json({ message: 'Password changed successfully' });
 	} catch (err) {
 		res.status(500).json({ message: err.message });
